bifurcated old and new code... (deleted old code, past versions will continue to exist in git)
