+++ config/config.go
@@ -3,11 +3,16 @@ package config
 import (
 	"github.com/joho/godotenv"
 	"log"
+	"os"
 )
 
 func Init() {
-	err := godotenv.Load()
-	if err != nil {
-		log.Fatal("Error loading .env file")
+	if os.Getenv("ENV") == "production" {
+		// 	Do production things
+	} else {
+		err := godotenv.Load()
+		if err != nil {
+			log.Fatal("Error loading .env file")
+		}
 	}
 }
