+++ submission/grader.go
@@ -1,4 +1,4 @@
-package submit
+package submission
 
 import (
 	"os/exec"
