--- submission/lint.go
@@ -1,65 +1,0 @@
-package submission
-
-import (
-	"context"
-	"fmt"
-	"os/exec"
-	"path/filepath"
-	"strings"
-	"time"
-)
-
-func (res *ResultSchema) LintCode(path string, submission SubmissionSchema) error {
-	absPath, _ := filepath.Abs(path)
-	cmd := exec.Command("grep", "-r", "int main()", "-H", ".")
-	cmd.Dir = absPath
-
-	grep, _ := cmd.CombinedOutput()
-
-	// [2:] basically removes ./ from the exec output ex: ./exec -> exec
-	lintData := LintThisCode(path, strings.Split(string(grep), ":")[0][2:])
-	lintMap := make(map[string]FileLint)
-	for _, line := range lintData {
-		tokens := strings.Split(line, ":")
-		if tokens[0] == "nofile" {
-			break
-		}
-		fileRef := findFile(tokens[0], submission.SubmissionFiles)
-		message := strings.Join(tokens[1:], ":")
-		lintMap[fileRef.ID] = FileLint{
-			FileID:    fileRef.ID,
-			LintLines: append(lintMap[fileRef.ID].LintLines, message),
-		}
-	}
-	for _, b := range lintMap {
-		res.SubmissionFileLint = append(res.SubmissionFileLint, b)
-	}
-
-	return nil
-}
-
-func findFile(name string, files []FileReference) FileReference {
-	for _, file := range files {
-		if name == file.FileName {
-			return file
-		}
-	}
-	return FileReference{}
-}
-
-func LintThisCode(path string, file string) []string {
-
-	var ctx, cancel = context.WithTimeout(context.Background(), time.Millisecond*1000)
-
-	absPath, _ := filepath.Abs(path)
-	cmd := exec.CommandContext(ctx, "bash", "-c", fmt.Sprintf("/opt/homebrew/Cellar/cppcheck/2."+
-		"3/bin/cppcheck --enable=all --inconclusive --library=posix --std=c++11 --quiet --template='{file}:{line"+
-		"}:{column}:{severity}:{message}' --language=c++ --cppcheck-build-dir=%s *.cpp *.h", absPath))
-	defer cancel()
-	cmd.Dir = absPath
-
-	_ = cmd.Wait()
-
-	mem, _ := cmd.CombinedOutput()
-	return strings.Split(string(mem), "\n")
-}
